Fix for reading generic data records.
The API for specific records and generic records were not
copatible, causing to generic data to simply be dropped.
This commit returns the generic data, while ignoring the
record passed in. Reading specific record continues to
populate the passed record.

This solves the immediate issue or dropping data. The API
is bit more clounky though.