Remove Tell() from the Encoder interface
This allows NewBinaryEncoder to take an io.Writer, which allows you to
satisfy it with `*bytes.Buffer` or any of a number of things,
including net.Conn and io.Pipe.

While this is technically a breaking change, I suspect nobody actually
is writing their own code which relies on Encoder, so it should be
fairly innocuous.